Refactor `ConfigAdjuster` for use from different test classes
